frontend/requirements.txt
# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# HTTP Requests
requests==2.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Async Support
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0

# Logging
structlog==24.1.0

# Optional: For future ML work
scikit-learn==1.4.0
scipy==1.12.0